@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 32
     defaultConfig {
         applicationId "fi.finwe.orion360.sdk.pro.examples"
         minSdkVersion 18
-        targetSdkVersion 30
-        versionCode 1101401
-        versionName "1.1.01.401"
+        targetSdkVersion 32
+        versionCode 1200400
+        versionName "1.2.00.40"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
@@ -27,7 +27,7 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.appcompat:appcompat:1.3.0'
+    implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.cardview:cardview:1.0.0'
     implementation files('libs/google-play-zip-file_lib.jar')
@@ -44,13 +44,18 @@ dependencies {
     //noinspection GradleDependency
     releaseImplementation 'fi.finwe.util:finwe-util-public:1.0.04'
 
-    // From Finwe maven repo at GitHub
+    // From Finwe maven repo at GitHub.
     implementation 'fi.finwe.orion360:orion360-sdk-pro-public:3.1.02.100'
 
     // ExoPlayer as an alternative video engine.
+    //noinspection GradleDependency
     implementation 'com.google.android.exoplayer:exoplayer-core:r2.4.1'
+    //noinspection GradleDependency
     implementation 'com.google.android.exoplayer:exoplayer-dash:r2.4.1'
+    //noinspection GradleDependency
     implementation 'com.google.android.exoplayer:exoplayer-hls:r2.4.1'
+    //noinspection GradleDependency
     implementation 'com.google.android.exoplayer:exoplayer-smoothstreaming:r2.4.1'
+    //noinspection GradleDependency
     implementation 'com.google.android.exoplayer:exoplayer-ui:r2.4.1'
 }
