@@ -35,18 +35,14 @@ import android.os.Bundle;
 import fi.finwe.orion360.sdk.pro.OrionActivity;
 import fi.finwe.orion360.sdk.pro.OrionScene;
 import fi.finwe.orion360.sdk.pro.OrionViewport;
-import fi.finwe.orion360.sdk.pro.controller.RotationAligner;
-import fi.finwe.orion360.sdk.pro.controller.TouchPincher;
-import fi.finwe.orion360.sdk.pro.controller.TouchRotater;
 import fi.finwe.orion360.sdk.pro.examples.MainMenu;
 import fi.finwe.orion360.sdk.pro.examples.R;
+import fi.finwe.orion360.sdk.pro.examples.TouchControllerWidget;
 import fi.finwe.orion360.sdk.pro.item.OrionCamera;
 import fi.finwe.orion360.sdk.pro.item.OrionPanorama;
 import fi.finwe.orion360.sdk.pro.item.OrionSceneItem;
 import fi.finwe.orion360.sdk.pro.source.OrionTexture;
 import fi.finwe.orion360.sdk.pro.view.OrionView;
-import fi.finwe.orion360.sdk.pro.widget.OrionWidget;
-import fi.finwe.util.ContextUtil;
 
 /**
  * An example of bindings for creating a player with an overview image on top.
@@ -120,7 +116,7 @@ public class Overview extends OrionActivity {
 
         // Create a new video (or image) texture from a video (or image) source URI.
         mPanoramaTexture = OrionTexture.createTextureFromURI(this,
-                MainMenu.PRIVATE_EXPANSION_FILES_PATH + MainMenu.TEST_VIDEO_FILE_MQ);
+                MainMenu.PRIVATE_ASSET_FILES_PATH + MainMenu.TEST_VIDEO_FILE_MQ);
 
         // Bind the panorama texture to the panorama objects. Here we assume full spherical
         // equirectangular monoscopic source, and wrap the complete texture around the sphere /
@@ -141,7 +137,7 @@ public class Overview extends OrionActivity {
         // Set yaw angle to 0. Now the camera will always point to the same angle
         // (to the center point of the equirectangular video/image source)
         // when starting the app, regardless of the orientation of the device.
-        mMainViewCamera.setRotationYaw(0.0f);
+        mMainViewCamera.setDefaultRotationYaw(0.0f);
 
         // Create a new camera for the overview viewport.
         mOverviewCamera = new OrionCamera();
@@ -149,13 +145,13 @@ public class Overview extends OrionActivity {
         // Set yaw angle to 0. Now the camera will always point to the same angle
         // (to the center point of the equirectangular video/image source)
         // when starting the app, regardless of the orientation of the device.
-        mOverviewCamera.setRotationYaw(0.0f);
+        mOverviewCamera.setDefaultRotationYaw(0.0f);
 
         // Bind camera as a controllable to sensor fusion. This will let sensors rotate the camera.
         mOrionContext.getSensorFusion().bindControllable(mMainViewCamera);
 
         // Create a new touch controller widget (convenience class) and let it control our camera.
-        mTouchController = new TouchControllerWidget(mMainViewCamera);
+        mTouchController = new TouchControllerWidget(mOrionContext, mMainViewCamera);
 
         // Bind the touch controller widget to the scene. This will make it functional in the scene.
         mSceneRectilinear.bindWidget(mTouchController);
@@ -167,8 +163,10 @@ public class Overview extends OrionActivity {
         // viewport per eye. Here we fill the complete view with one (landscape) viewport, and
         // add another much smaller one on top of it to function as an overview image.
         mView.bindViewports(new RectF[] {
-                new RectF(0.0F, 1.0F, 1.0F, 0.0F),          // Main covers the whole view
-                new RectF(0.35F, 0.25F, 0.65F, 0.0F) },     // Overview covers small bottom area
+                new RectF(0.0F, 1.0F, 1.0F, 0.0F),
+                // above: Main covers the whole view
+                new RectF(0.355F, 0.25F, 0.645F, 0.0F) },
+                // above: Overview covers small bottom area
                 OrionViewport.CoordinateType.FIXED_LANDSCAPE);
 
         // Notice the viewport rect coordinate system. The viewport coordinates are relative
@@ -192,69 +190,4 @@ public class Overview extends OrionActivity {
         // Bind the overview viewport to the overview viewport camera.
         mView.getViewports()[1].bindCamera(mOverviewCamera);
 	}
-
-    /**
-     * Convenience class for configuring typical touch control logic.
-     */
-    public class TouchControllerWidget implements OrionWidget {
-
-        /** The camera that will be controlled by this widget. */
-        private OrionCamera mCamera;
-
-        /** Touch pinch-to-zoom/pinch-to-rotate gesture handler. */
-        private TouchPincher mTouchPincher;
-
-        /** Touch drag-to-pan gesture handler. */
-        private TouchRotater mTouchRotater;
-
-        /** Rotation aligner keeps the horizon straight at all times. */
-        private RotationAligner mRotationAligner;
-
-
-        /**
-         * Constructs the widget.
-         *
-         * @param camera The camera to be controlled by this widget.
-         */
-        TouchControllerWidget(OrionCamera camera) {
-
-            // Keep a reference to the camera that we control.
-            mCamera = camera;
-
-            // Create pinch-to-zoom/pinch-to-rotate handler.
-            mTouchPincher = new TouchPincher();
-            mTouchPincher.setMinimumDistanceDp(mOrionContext.getActivity(), 20);
-            mTouchPincher.bindControllable(mCamera, OrionCamera.VAR_FLOAT1_ZOOM);
-
-            // Create drag-to-pan handler.
-            mTouchRotater = new TouchRotater();
-            mTouchRotater.bindControllable(mCamera);
-
-            // Create the rotation aligner, responsible for rotating the view so that the horizon
-            // aligns with the user's real-life horizon when the user is not looking up or down.
-            mRotationAligner = new RotationAligner();
-            mRotationAligner.setDeviceAlignZ(-ContextUtil.getDisplayRotationDegreesFromNatural(
-                    mOrionContext.getActivity()));
-            mRotationAligner.bindControllable(mCamera);
-
-            // Rotation aligner needs sensor fusion data in order to do its job.
-            mOrionContext.getSensorFusion().bindControllable(mRotationAligner);
-        }
-
-        @Override
-        public void onBindWidget(OrionScene scene) {
-            // When widget is bound to scene, bind the controllers to it to make them functional.
-            scene.bindController(mTouchPincher);
-            scene.bindController(mTouchRotater);
-            scene.bindController(mRotationAligner);
-        }
-
-        @Override
-        public void onReleaseWidget(OrionScene scene) {
-            // When widget is released from scene, release the controllers as well.
-            scene.releaseController(mTouchPincher);
-            scene.releaseController(mTouchRotater);
-            scene.releaseController(mRotationAligner);
-        }
-    }
 }
