@@ -32,7 +32,6 @@ package fi.finwe.orion360.sdk.pro.examples.binding;
 import android.graphics.RectF;
 import android.os.Bundle;
 
-import fi.finwe.math.Quatf;
 import fi.finwe.orion360.sdk.pro.OrionActivity;
 import fi.finwe.orion360.sdk.pro.OrionScene;
 import fi.finwe.orion360.sdk.pro.OrionViewport;
@@ -71,11 +70,11 @@ public class RearviewMirror extends OrionActivity {
     /** The Android view where our 3D scene will be rendered to. */
     protected OrionView mView;
 
-    /** The 3D scene where our panorama sphere will be added to. */
+    /** The 3D scenes where our panorama spheres will be added to. */
     protected OrionScene mScene;
     protected OrionScene mSceneRearview;
 
-    /** The panorama sphere where our video texture will be mapped to. */
+    /** The panorama spheres where our video textures will be mapped to. */
     protected OrionPanorama mPanorama;
     protected OrionPanorama mPanoramaRearview;
 
@@ -97,46 +96,42 @@ public class RearviewMirror extends OrionActivity {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_main);
 
-        // Create a new scene. This represents a 3D world where various objects can be placed.
+        // Create new scenes. These represent 3D worlds where various objects can be placed.
         mScene = new OrionScene();
         mSceneRearview = new OrionScene();
 
         // Bind sensor fusion as a controller. This will make it available for scene objects.
         mScene.bindController(mOrionContext.getSensorFusion());
 
-        // Create a new panorama. This is a 3D object that will represent a spherical video/image.
+        // Create new panoramas. These are 3D objects that will represent a spherical video/image.
         mPanorama = new OrionPanorama();
         mPanoramaRearview = new OrionPanorama();
 
         // Create a new video (or image) texture from a video (or image) source URI.
         mPanoramaTexture = OrionTexture.createTextureFromURI(this,
-                MainMenu.PRIVATE_EXPANSION_FILES_PATH + MainMenu.TEST_VIDEO_FILE_MQ);
+                MainMenu.PRIVATE_ASSET_FILES_PATH + MainMenu.TEST_VIDEO_FILE_MQ);
 
-        // Bind the panorama texture to the panorama object. Here we assume full spherical
+        // Bind the panorama textures to the panorama objects. Here we assume full spherical
         // equirectangular monoscopic source, and wrap the complete texture around the sphere.
         // If you have stereoscopic content or doughnut shape video, use other method variants.
         mPanorama.bindTextureFull(0, mPanoramaTexture);
         mPanoramaRearview.bindTextureFull(0, mPanoramaTexture);
 
-        // Bind the panorama to the scene. This will make it part of our 3D world.
+        // Bind the panoramas to the scenes. This will make them part of our 3D worlds.
         mScene.bindSceneItem(mPanorama);
         mSceneRearview.bindSceneItem(mPanoramaRearview);
 
         // Create a new camera (main view). This will become the end-user's eyes into the 3D world.
         mMainViewCamera = new OrionCamera();
 
-        // Set yaw angle to 0. Now the camera will always point to the same angle
-        // (to the center point of the equirectangular video/image source)
-        // when starting the app, regardless of the orientation of the device.
-        mMainViewCamera.setRotation(Quatf.fromRotationAxisY(0.0 / Quatf.RAD));
+        // Reset view to the 'front' direction (horizontal center of the panorama).
+        mMainViewCamera.setDefaultRotationYaw(0);
 
         // Create a new camera (rear-view). This will become the end-user's eyes into the 3D world.
         mRearViewCamera = new OrionCamera();
 
-        // Set yaw angle to -180. Now the camera will always point to the same yaw angle
-        // (to the horizontal left edge of the equirectangular video/image source)
-        // when starting the app, regardless of the orientation of the device.
-        mRearViewCamera.setRotation(Quatf.fromRotationAxisY(180.0f / Quatf.RAD));
+        // Reset view to the 'back' direction.
+        mRearViewCamera.setDefaultRotationYaw(180.0f);
 
         // Bind camera as a controllable to sensor fusion. This will let sensors rotate the camera.
         mOrionContext.getSensorFusion().bindControllable(mMainViewCamera);
@@ -145,6 +140,7 @@ public class RearviewMirror extends OrionActivity {
         mOrionContext.getSensorFusion().bindControllable(mRearViewCamera);
 
         // Create a new touch controller widget (convenience class) and let it control our cameras.
+        // This is a slightly modified version compared to the usual one, with two cameras.
         mTouchController = new TouchControllerWidget(mMainViewCamera, mRearViewCamera);
 
         // Bind the touch controller widget to the scene. This will make it functional in the scene.
@@ -153,15 +149,14 @@ public class RearviewMirror extends OrionActivity {
         // Find Orion360 view from the XML layout. This is an Android view where we render content.
         mView = (OrionView)findViewById(R.id.orion_view);
 
-        // Bind the scene to the view. This is the 3D world that we will be rendering to this view.
-        //mView.bindDefaultScene(mGalleryScene);
-
         // The view can be divided into one or more viewports. For example, in VR mode we have one
         // viewport per eye. Here we fill the complete view with one (landscape) viewport, and
         // add another much smaller one on top of it to function as a rear-view mirror.
         mView.bindViewports(new RectF[] {
-                new RectF(0.0F, 1.0F, 1.0F, 0.0F),          // Main view covers the whole view
-                new RectF(0.25F, 0.95F, 0.75F, 0.75F) },    // Rear-view covers small area on top
+                new RectF(0.0F, 1.0F, 1.0F, 0.0F),
+                // above: Main view covers the whole view
+                new RectF(0.25F, 0.95F, 0.75F, 0.75F) },
+                // above: Rear-view covers small area on top
                 OrionViewport.CoordinateType.FIXED_LANDSCAPE);
 
         // Notice the viewport rect coordinate system. The viewport coordinates are relative
@@ -187,20 +182,14 @@ public class RearviewMirror extends OrionActivity {
      */
     public class TouchControllerWidget implements OrionWidget {
 
-        /** The main camera that will be controlled by this widget. */
-        private OrionCamera mMainCamera;
-
-        /** The rear-view camera that will be controlled by this widget. */
-        private OrionCamera mRearViewCamera;
-
         /** Touch pinch-to-zoom/pinch-to-rotate gesture handler. */
-        private TouchPincher mTouchPincher;
+        private final TouchPincher mTouchPincher;
 
         /** Touch drag-to-pan gesture handler. */
-        private TouchRotater mTouchRotater;
+        private final TouchRotater mTouchRotater;
 
         /** Rotation aligner keeps the horizon straight at all times. */
-        private RotationAligner mRotationAligner;
+        private final RotationAligner mRotationAligner;
 
 
         /**
@@ -211,40 +200,32 @@ public class RearviewMirror extends OrionActivity {
          */
         TouchControllerWidget(OrionCamera mainCamera, OrionCamera rearViewCamera) {
 
-            // Keep references to the cameras that we control.
-            mMainCamera = mainCamera;
-            mRearViewCamera = rearViewCamera;
-
             // Create pinch-to-zoom/pinch-to-rotate handler for the main camera only.
             // Notice that we do not want zooming to rear-view mirror, hence no binding.
             mTouchPincher = new TouchPincher();
             mTouchPincher.setMinimumDistanceDp(mOrionContext.getActivity(), 20);
-            mTouchPincher.bindControllable(mMainCamera, OrionCamera.VAR_FLOAT1_ZOOM);
+            mTouchPincher.bindControllable(mainCamera, OrionCamera.VAR_FLOAT1_ZOOM);
 
             // Create drag-to-pan handler.
             // Notice that we want panning to affect to both main and rear-view cameras.
             mTouchRotater = new TouchRotater();
-            mTouchRotater.bindControllable(mMainCamera);
-            mTouchRotater.bindControllable(mRearViewCamera);
+            mTouchRotater.bindControllable(mainCamera);
+            mTouchRotater.bindControllable(rearViewCamera);
 
             // Create the rotation aligner, responsible for rotating the view so that the horizon
             // aligns with the user's real-life horizon when the user is not looking up or down.
             mRotationAligner = new RotationAligner();
             mRotationAligner.setDeviceAlignZ(-ContextUtil.getDisplayRotationDegreesFromNatural(
                     mOrionContext.getActivity()));
-            mRotationAligner.bindControllable(mMainCamera);
-            mRotationAligner.bindControllable(mRearViewCamera);
+            mRotationAligner.bindControllable(mainCamera);
+            mRotationAligner.bindControllable(rearViewCamera);
 
             // Rotation aligner needs sensor fusion data in order to do its job.
             mOrionContext.getSensorFusion().bindControllable(mRotationAligner);
 
             // Notice that touch gestures will work when performed either on main viewport or
-            // rear-view viewport, but if user has zoomed in the amount of panning applied is
-            // relative to the zoom level of the viewport where the gesture is performed.
-            // As an example, consider a case where the main view is zoomed in. Now, panning
-            // the main view is slow (follows finger movement on main view) but panning the
-            // rear-view mirror is fast (follows finger movement on rear-view mirror).
-            // Tip: you can create quick pan control stripes with this technique.
+            // rear-view viewport. If user has zoomed in the amount of panning applied is
+            // relative to the zoom level.
         }
 
         @Override
